Implement hooks for workouts component

+++ gym-class-heroes/src/pages/WorkoutsPage.tsx
@@ -2,51 +2,84 @@ import { useEffect, useState } from 'react';
 import Workouts from "../components/workouts/workouts";
 import type { WorkoutsInterface } from '../components/interface/workoutsInterface';
 import WorkoutForm from '../components/form/workoutForm';
-import type { GroupArrayKey } from '../components/interface/groupArrayKey';
-import type { GroupsInterface } from '../components/interface/groupsInterface';
 import * as workoutService from "../services/workoutServices";
+import { useGroupData } from '../hooks/useGroupData';
+import { workoutData } from '../data/workoutData';
 
-type WorkoutPageProps = {
-    groupsData: GroupsInterface[];
-    addToGroup: (groupId: string, key: GroupArrayKey, subjectId: number) => void;
-    removeFromGroup: (groupId: string, key: GroupArrayKey, subjectId: number) => void; 
-}
-export default function WorkoutsPage({groupsData, addToGroup, removeFromGroup}: WorkoutPageProps) {
-    const [workouts, setWorkouts] = useState<WorkoutsInterface[]>([]);
+export default function WorkoutsPage() {
+    const [workouts, setWorkouts] = useState<WorkoutsInterface[]>(workoutData);
+
+    const {
+        groups,
+        error,
+        addToGroup,
+        removeFromGroup        
+    } = useGroupData();
 
     useEffect(() => {
-        const workoutData = workoutService.fetchWorkouts();
-        setWorkouts(workoutData);
+        const fetchWorkouts = async () => {
+            const workouts =
+                await workoutService.fetchWorkouts();
+
+            setWorkouts([...workouts]);
+        }; 
+        fetchWorkouts();
     }, []);
 
-    const onAddWorkout = (newWorkout: WorkoutsInterface) => {
-        const workoutResult = workoutService.createWorkout(newWorkout);
+    const onAddWorkout = async (newWorkout: WorkoutsInterface) => {
+        try {
 
-        if (typeof workoutResult === "string") {
-            console.error(workoutResult);
+        const createWorkout = await workoutService.createWorkout(newWorkout);
+
+        if (typeof createWorkout === "string") {
+            console.error(createWorkout);
             return;
         }
 
-        setWorkouts(workoutService.fetchWorkouts());
+        setWorkouts(prev => [
+            ...prev,
+            createWorkout
+        ]);
 
-        addToGroup(workoutResult.group, "workoutsById", workoutResult.id);
+        addToGroup(newWorkout.group, "workoutsById", newWorkout.id);
+
+        } catch (error) {
+            console.error(error);
+        }
     };
 
-    const onRemoveWorkout = (workoutId: number) => {
-        const workout = workoutService.getWorkoutById(workoutId);
+    const onRemoveWorkout = async (
+        workoutId: number
+    ) => {
+        try {
 
-        workoutService.deleteWorkout(workoutId);
-        
-        setWorkouts(workoutService.fetchWorkouts());
+        const workout =
+            workouts.find(w => w.id === workoutId);
 
-        removeFromGroup(workout.group, "workoutsById", workoutId);
+        if (!workout) return;
+
+        await workoutService.deleteWorkout(workoutId);
+
+        removeFromGroup(
+            workout.group,
+            "workoutsById",
+            workoutId
+        );
+
+        setWorkouts(prev =>
+            prev.filter(w => w.id !== workoutId)
+        );
+
+        } catch (error) {
+            console.error(error);
+        }
     };
 
     return (
         <>
             <div className='flex flex-col w-full min-h-screen bg-gray-50'>
                 <Workouts workouts={workouts} onRemoveWorkout={onRemoveWorkout}/>
-                <WorkoutForm groupsData={groupsData} onAddWorkout={onAddWorkout}/>
+                <WorkoutForm groupsData={groups} onAddWorkout={onAddWorkout}/>
             </div>
             
         </>
